Add first pictures endpoints

+++ backend/src/main/java/be/uliege/speam/team03/MDTools/repositories/CategoryRepository.java
@@ -11,16 +11,16 @@ import be.uliege.speam.team03.MDTools.models.Category;
 import be.uliege.speam.team03.MDTools.models.SubGroup;
 
 @Repository
-public interface CategoryRepository extends CrudRepository<Category, Integer> {
+public interface CategoryRepository extends CrudRepository<Category, Long> {
     @Query("SELECT cc.val FROM CategoryCharacteristic cc " +
            "JOIN cc.characteristic c " +
            "WHERE cc.category.id = :categoryId AND c.name = :characteristicName")
-    Optional<String> findCharacteristicVal(Integer categoryId, String characteristicName);
+    Optional<String> findCharacteristicVal(Long categoryId, String characteristicName);
 
     @Query("SELECT cc.valAbrev FROM CategoryCharacteristic cc " +
            "JOIN cc.characteristic c " +
            "WHERE cc.category.id = :categoryId AND c.name = :characteristicName")
-    Optional<String> findCharacteristicValAbrv(Integer categoryId, String characteristicName);
+    Optional<String> findCharacteristicValAbrv(Long categoryId, String characteristicName);
     
     Optional<List<Category>> findBySubGroup(SubGroup subGroup);
     Optional<List<Category>> findBySubGroupIn(List<SubGroup> subGroups);
